gedit-plugins: don't build the charmap plugin
The unicode-character-database module doesn't build anymore, the Flathub
bot gives some errors.

In gedit-plugins 48.2 (here we stay on 48.1 for now) the charmap plugin
has anyway been removed, because it's written in Python and Python
plugins are all broken with the latest pygobject version.

So, instead of trying to fix the build of the charmap deps, don't build
charmap at all and remove the deps. That way, when moving to
gedit >= 48.2 it will be one less task to do.